add api gateway v1 integration
