+++ packages/remix-architect/api/v2.ts
@@ -1,0 +1,106 @@
+import type {
+  APIGatewayProxyEventHeaders,
+  APIGatewayProxyEventV2,
+  APIGatewayProxyStructuredResultV2
+} from "aws-lambda";
+import {
+  Headers as NodeHeaders,
+  Request as NodeRequest
+} from "@remix-run/node";
+import type {
+  // This has been added as a global in node 15+
+  AbortController,
+  Response as NodeResponse,
+} from "@remix-run/node";
+
+import { isBinaryType } from "../binaryTypes";
+
+export function createRemixRequest(
+  event: APIGatewayProxyEventV2,
+  abortController?: AbortController
+): NodeRequest {
+  let host = event.headers["x-forwarded-host"] || event.headers.host;
+  let search = event.rawQueryString.length ? `?${event.rawQueryString}` : "";
+  let scheme = process.env.ARC_SANDBOX ? "http" : "https";
+  let url = new URL(event.rawPath + search, `${scheme}://${host}`);
+  let isFormData = event.headers["content-type"]?.includes(
+    "multipart/form-data"
+  );
+
+  return new NodeRequest(url.href, {
+    method: event.requestContext.http.method,
+    headers: createRemixHeaders(event.headers, event.cookies),
+    body:
+      event.body && event.isBase64Encoded
+        ? isFormData
+          ? Buffer.from(event.body, "base64")
+          : Buffer.from(event.body, "base64").toString()
+        : event.body,
+    abortController,
+    signal: abortController?.signal,
+  });
+}
+
+export function createRemixHeaders(
+  requestHeaders: APIGatewayProxyEventHeaders,
+  requestCookies?: string[]
+): NodeHeaders {
+  let headers = new NodeHeaders();
+
+  for (let [header, value] of Object.entries(requestHeaders)) {
+    if (value) {
+      headers.append(header, value);
+    }
+  }
+
+  if (requestCookies) {
+    headers.append("Cookie", requestCookies.join("; "));
+  }
+
+  return headers;
+}
+
+export async function sendRemixResponse(
+  nodeResponse: NodeResponse,
+  abortController: AbortController
+): Promise<APIGatewayProxyStructuredResultV2> {
+  let cookies: string[] = [];
+
+  // Arc/AWS API Gateway will send back set-cookies outside of response headers.
+  for (let [key, values] of Object.entries(nodeResponse.headers.raw())) {
+    if (key.toLowerCase() === "set-cookie") {
+      for (let value of values) {
+        cookies.push(value);
+      }
+    }
+  }
+
+  if (cookies.length) {
+    nodeResponse.headers.delete("Set-Cookie");
+  }
+
+  if (abortController.signal.aborted) {
+    nodeResponse.headers.set("Connection", "close");
+  }
+
+  let contentType = nodeResponse.headers.get("Content-Type");
+  let isBinary = isBinaryType(contentType);
+  let body;
+  let isBase64Encoded = false;
+
+  if (isBinary) {
+    let blob = await nodeResponse.arrayBuffer();
+    body = Buffer.from(blob).toString("base64");
+    isBase64Encoded = true;
+  } else {
+    body = await nodeResponse.text();
+  }
+
+  return {
+    statusCode: nodeResponse.status,
+    headers: Object.fromEntries(nodeResponse.headers),
+    cookies,
+    body,
+    isBase64Encoded,
+  };
+}
