+++ packages/remix-architect/server.ts
@@ -1,15 +1,13 @@
 import {
   // This has been added as a global in node 15+
   AbortController,
-  Headers as NodeHeaders,
-  Request as NodeRequest,
   createRequestHandler as createRemixRequestHandler,
 } from "@remix-run/node";
 import type {
-  APIGatewayProxyEventHeaders,
+  APIGatewayProxyEvent,
   APIGatewayProxyEventV2,
+  APIGatewayProxyHandler,
   APIGatewayProxyHandlerV2,
-  APIGatewayProxyStructuredResultV2,
 } from "aws-lambda";
 import type {
   AppLoadContext,
@@ -17,7 +15,14 @@ import type {
   Response as NodeResponse,
 } from "@remix-run/node";
 
-import { isBinaryType } from "./binaryTypes";
+import {
+  sendRemixResponse as sendRemixResponseV2,
+  createRemixRequest as createRemixRequestV2
+} from "./api/v2";
+import {
+  createRemixRequest,
+  sendRemixResponse
+} from "./api/v1";
 
 /**
  * A function that returns the value to use as `context` in route `loader` and
@@ -27,10 +32,15 @@ import { isBinaryType } from "./binaryTypes";
  * environment/platform-specific values through to your loader/action.
  */
 export type GetLoadContextFunction = (
-  event: APIGatewayProxyEventV2
+  event: APIGatewayProxyEventV2 | APIGatewayProxyEvent
 ) => AppLoadContext;
 
-export type RequestHandler = APIGatewayProxyHandlerV2;
+export type RequestHandler = APIGatewayProxyHandlerV2 | APIGatewayProxyHandler;
+
+export enum APIGatewayVersion {
+  v1 = "v1",
+  v2 = "v2",
+}
 
 /**
  * Returns a request handler for Architect that serves the response using
@@ -40,16 +50,20 @@ export function createRequestHandler({
   build,
   getLoadContext,
   mode = process.env.NODE_ENV,
+  apiGatewayVersion = APIGatewayVersion.v2
 }: {
   build: ServerBuild;
   getLoadContext?: GetLoadContextFunction;
   mode?: string;
+  apiGatewayVersion?: APIGatewayVersion;
 }): RequestHandler {
   let handleRequest = createRemixRequestHandler(build, mode);
 
-  return async (event /*, context*/) => {
+  return async (event: APIGatewayProxyEvent | APIGatewayProxyEventV2 /*, context*/) => {
     let abortController = new AbortController();
-    let request = createRemixRequest(event, abortController);
+    let request = apiGatewayVersion === APIGatewayVersion.v1
+      ? createRemixRequest(event as APIGatewayProxyEvent, abortController)
+      : createRemixRequestV2(event as APIGatewayProxyEventV2, abortController);
     let loadContext =
       typeof getLoadContext === "function" ? getLoadContext(event) : undefined;
 
@@ -58,96 +72,8 @@ export function createRequestHandler({
       loadContext
     )) as unknown as NodeResponse;
 
-    return sendRemixResponse(response, abortController);
-  };
-}
-
-export function createRemixRequest(
-  event: APIGatewayProxyEventV2,
-  abortController?: AbortController
-): NodeRequest {
-  let host = event.headers["x-forwarded-host"] || event.headers.host;
-  let search = event.rawQueryString.length ? `?${event.rawQueryString}` : "";
-  let scheme = process.env.ARC_SANDBOX ? "http" : "https";
-  let url = new URL(event.rawPath + search, `${scheme}://${host}`);
-  let isFormData = event.headers["content-type"]?.includes(
-    "multipart/form-data"
-  );
-
-  return new NodeRequest(url.href, {
-    method: event.requestContext.http.method,
-    headers: createRemixHeaders(event.headers, event.cookies),
-    body:
-      event.body && event.isBase64Encoded
-        ? isFormData
-          ? Buffer.from(event.body, "base64")
-          : Buffer.from(event.body, "base64").toString()
-        : event.body,
-    abortController,
-    signal: abortController?.signal,
-  });
-}
-
-export function createRemixHeaders(
-  requestHeaders: APIGatewayProxyEventHeaders,
-  requestCookies?: string[]
-): NodeHeaders {
-  let headers = new NodeHeaders();
-
-  for (let [header, value] of Object.entries(requestHeaders)) {
-    if (value) {
-      headers.append(header, value);
-    }
-  }
-
-  if (requestCookies) {
-    headers.append("Cookie", requestCookies.join("; "));
-  }
-
-  return headers;
-}
-
-export async function sendRemixResponse(
-  nodeResponse: NodeResponse,
-  abortController: AbortController
-): Promise<APIGatewayProxyStructuredResultV2> {
-  let cookies: string[] = [];
-
-  // Arc/AWS API Gateway will send back set-cookies outside of response headers.
-  for (let [key, values] of Object.entries(nodeResponse.headers.raw())) {
-    if (key.toLowerCase() === "set-cookie") {
-      for (let value of values) {
-        cookies.push(value);
-      }
-    }
-  }
-
-  if (cookies.length) {
-    nodeResponse.headers.delete("Set-Cookie");
-  }
-
-  if (abortController.signal.aborted) {
-    nodeResponse.headers.set("Connection", "close");
-  }
-
-  let contentType = nodeResponse.headers.get("Content-Type");
-  let isBinary = isBinaryType(contentType);
-  let body;
-  let isBase64Encoded = false;
-
-  if (isBinary) {
-    let blob = await nodeResponse.arrayBuffer();
-    body = Buffer.from(blob).toString("base64");
-    isBase64Encoded = true;
-  } else {
-    body = await nodeResponse.text();
-  }
-
-  return {
-    statusCode: nodeResponse.status,
-    headers: Object.fromEntries(nodeResponse.headers),
-    cookies,
-    body,
-    isBase64Encoded,
+    return apiGatewayVersion === APIGatewayVersion.v1
+        ? sendRemixResponse(response, abortController)
+        : sendRemixResponseV2(response, abortController);
   };
 }
